+++ flutter-hms-push/android/src/main/java/com/huawei/hms/flutter/push/receiver/common/NotificationIntentListener.java
@@ -48,7 +48,7 @@ public class NotificationIntentListener implements NewIntentListener {
 
     private String initialIntent;
 
-    private Context context;
+    private final Context context;
 
     public NotificationIntentListener(Context context) {
         this.context = context;